backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
python-dotenv
openai
pillow
google-genai
pydantic>=2.0.0
python-dotenv
google-cloud-vision
google-cloud-storagepypdf==6.6.2
